backend/requirements.txt
fastapi
sqlalchemy
uvicorn
databases
pydantic
pydantic[email]
passlib[bcrypt]
python-multipart
python-socketio
python-jose[cryptography]
aiofiles
requests
bcrypt